bugfix: escape assistant messages when rendering is off in replay function
Note, the code changes are made by `debgpt --inplace ...`.
The original user instruction is: 'there is a bug. when render is off, the assistant message is not rendered with markdown, but rich will interpret `[]` and make printing incorrect. please use escape to fix that issue. Just show me the result and do not say anything else. No need to enclose the result using "```".'

Note, this commit message is generated by debgpt.